rename avatarConfig file
