+++ react/features/base/avatar/components/web/avatarConfig.js
